Fix "Verbose" option to "check" after recent capture changes …
Previously, we computed the filename of the temporary file containing
the output of failing tests based on the number of the test itself in
relation to the total number of tests being run.  However, now that
"capture" is used for testing, this will no longer work, as temporary
files are not always created, and the temporary filename counter will
not be incremented in a predictable manner.

Instead, we store the location of the temporary file immediately after
the test is run.  In particular, the local variable "errorList" now
contains tuples of the form "(test number, temporary file number)".