Check package using tests from package test directory, if it exists
This is currently only implemented for Core, as it is the only package
that has a directory for its tests.  However, the helper function
"loadTestDir" will work for any packages that might install tests in
such a directory in the future.

We add a new "test directory loaded" key to the package hash table.
If it is not present, then we have not tried to load the tests.  If it
is present but false, then we tried, but the directory does not exist.
If it is present but true, then we tried and were successful.