Fix in-source autotools installation of Core test files
When building in-source, the "SRCDIR" variable is just ".", and so
"subst" was substituting "@pre_docdir@/Core/tests" at each dot,
including the one in ".m2" at the end of the filename.  This was
causing "No rule to make target" build failures.

We use "patsubst" instead and are more particular about what we are
substituting.