fix the way "setup()" handles INFOPATH
old way:

$ unset INFOPATH
$ . .profile-Macaulay2
$ printenv INFOPATH
/Users/dan/src/M2/M2.git/M2/BUILD/dan/builds.tmp/-development/usr-dist/common/share/info

new way:

$ unset INFOPATH
$ . .profile-Macaulay2
$ printenv INFOPATH
/Users/dan/src/M2/M2.git/M2/BUILD/dan/builds.tmp/-development/usr-dist/common/share/info:

The point is that the empty entry at the end of the path tells emacs-info to search in the standard places, too.
Here is what those standard places are on my Mac:

    /Applications/Emacs-27.1-1.app/Contents/Resources/info
    /Applications/Emacs-27.1-1.app/Contents/Resources/info/
    /usr/local/share/info/
    /usr/share/info/

Our users must have been missing out on a lot of info files for 15 years.