Fix line number of tests loaded from files
Previously, we used line 1 (the beginning of the code), but this isn't
consistent with tests loaded from strings (which use the line at the
end of the code).  We make the two cases consistent.  In particular,
this fixes (locate, TestInput) for tests loaded from files (e.g., the
Core tests).