a fix for github issue #2051.  A test was added to normal/inverse.m2
The problem was that an assumption I made in computing inverses in rings did not hold over the rationals, due to denominators.  The fix involves multiplying by the correct rational number (actually integer) to correct the problem.

Another bug in experimental code (not used yet in production) was also fixed, but that should not affect anyone's code.