Restore pythonObjectCell case in tostringfun
It was removed to make way for implementing `toString PythonObject` at
top level in a43a9a5.  To avoid compilation errors, an else clause was
added to the end of `tostringfun`.

However, this was not great, as now it's more likely that anyone
wishing to add a new member to the `Expr` union would miss adding
their new case to this function.

Instead, we restore a default version of the `pythonObjectCell` case
that will be overridden at top level.