Replace spaces in wikipedia links with underscores
This gives the canonical link to articles with spaces in their titles
(see https://en.wikipedia.org/wiki/Wikipedia:Linking_to_Wikipedia).

This way, when using wikipedia(String), we can give more
human-readable names to articles (with a space), but the actual link
will use an underscore.  This is helpful, for example,  for viewing
the documentation in Emacs while using goto-address-mode to get clickable
links.  Previously, only the part of the url before a space would have
been clickable.