Use AC_ARG_WITH instead of AC_ARG_ENABLE for mathicgb/etc. checks
It's a subtle difference, but AC_ARG_ENABLE is intended for use with
internal features and AC_ARG_WITH for external packages, so the latter
is more appropriate in this case.

So to use the mathicgb library already available on the system, we now
use --with-system-mathicgb instead of --disable-build-mathicgb.

This also adds consistency with the new --with-system-gc option.