Add helper function for demarking values in JSON arrays and objects
Reduces code duplication since they're essentially the same case.
Note that we pass a function instead of a list of values to demark so
that we don't mess up the indentation.  Otherwise, we'll compute the
indentation inside the values before we compute it between the values,
but we need it to be the other way around.