Properly allocate memory for ForeignFunction return value
I had left 100 bytes in as a dummy value during initial testing but
neglected to go back and fix it until now.  We either allocate the
memory needed to store the type of the return value, or if it's a
smaller type, widen that to the system register size (libffi has a
special typedef for this, ffi_arg, but we just use version#"pointer
size").

We also simplify on line of corresponding code in the d directory,
using the helper function "getMem" instead of calling "Ccode" directly.