Polyhedra: Fix bug in regularSubdivision(Matrix, Matrix)
In this method a lower envelope is computed whose (projected) facets in
turn are a regular subdivision of the input point set. Previously
the return values probably were actual polyhedra, but now these are just
sets of indices indicating which points together form a cell. However
these indices use the vertices of the intermediate polyhedron used for
computing the lower envelope. Thus they need to be remapped onto the
input points.