Add "packages" button to top of html documentation pages
This allows users to easily switch from viewing a documentation page
in one package to viewing a document page for another package.

We replace some commented out code referring to a "directoryButton",
which may have been intended to point to the directory page that is
installed in applicationDirectory.  However, the URL to this page
would be impossible to know when the documentation is built for a
distribution because it depends on the value of the user's HOME
environment variable and the applicationDirectorySuffix variable.  The
"packages provided with Macaulay2" page in Macaulay2Doc is the next
best thing.