Update autotools build to build prism syntax highlighting library
It is not part of the regular build, but is run prior to each release
by running:

    make -C Macaulay2/editors update-syntax-highlighting

This builds prism.js, copies it to the Style package, and commits the
changes to git.

We can also still build hightlight.js manually.

Note that we also remove @srcdir@ from VPATH as we want to make sure
that various prerequisite files get copied to the build directory.