Add GitHub workflow for validating HTML documentation
Only runs on Ubuntu autotools build.  On a pull request, if any
changes to Core (and therefore, possibly the generation of HTML
documentation), we validate everything.  Otherwise, we only validate
the documentation to the packages that were changed.

On no-pull requests (i.e., scheduled builds or pushes to pre-master),
validate everything.