`LeftIdeal` replaces the old `Ideal`; new `Ideal` is two-sided
Ideal is a type of LeftIdeal.
Most functions are defined for LeftIdeal; the ones that make sense
only for a two-sided ideal are defined for Ideal.
The method `ideal` constructs an ideal typical for the given ring:
e.g., LeftIdeal for a Weyl algebra and Ideal for a commutative or
skew-commutative ring.

Current state:
FirstPackage, Style, Macaulay2Doc install without errors,
there are errors in packages related to D-modules.