Improve error message for toExternalString(Function)
When passed an anonymous function with no global name, it prints the
location of the function, but enclosed in parentheses.  The Emacs
regular expression thinks that the parentheses are part of the
filename, so compilation mode fails to jump to the right place and a
file dialog opens.

So we call net(Function) to enclose it in the usual Emacs-friendly
"-*Function[...]*-".  We also drop the call the "locate", since "net"
will do that for us.