Add new evaluatedCode member of Code union
This will be used by augmented assignment.  The problem was that we
first want to check if a user-defined method for augmented assignment
is installed, which requires evaluating the code on the left-hand side
of the operator.  But if we *don't* use a user-defined method (which
will be the case most of the time), then we want the original code,
since that's what the binary operators expect.

The solution is to create a member of the Code union that saves the
evaluated code.  Then we can evaluate early on but still call the
binary operators later if need.