Move the check for evaluating tryCode to its own helper function
This way we can use it for null coalescion as well.  We pass a thread
local variable with a pointer that will contain the result of the
evaluation because we want the helper function to return a
boolean (whether we return the evaluated code or move to the next
thing).