correction to RealRoots:-realRootIsolation
Fixed oversight in realRootIsolation method that fails to perform bisection if the polynomial's leadCoefficient is negative. Lagrange bound was incorrectly implemented, failing to return a positive values whenever the lead coefficient was negative, and included the leading coefficient. This should now be correct (I have tested this but would appreciate another pair of eyes on this just in case).