Further improved bound and skipped bisection for isolated roots within interval size
Combined the Lagrange and Cauchy bounds to return the smallest. realRootIsolation no longer bisects intervals when the root is isolated and the interval is below the required size (e.g. when two roots are very close together, the interval widths with be smaller, but the interval widths do not need to be smaller for the other roots). Final list is also sorted in case bisection stops in an interval after the first.

Test also corrected (the new isolating intervals only look 'less nice' because the previous initial bounds happened to be divisible by 8).