Further corrections/improvements to RealRoots:-realRootIsolation
Correction to some issues that arose (if polynomial was of the form Ax^2-Bx, Lagrange bound would miss the root |x|=|B/A|; the case where the polynomial is one term (i.e Ax^n) was improved by simply setting the bound to 0 to avoid any further calls to bisection algorithm). Actual comparisons between Cauchy and Lagrange bound revealed if the Lagrange bound is ever better, is it better by at most one, so Lagrange bound has been removed (this also avoids roots being missed in the above case).

Knuth bound was also introduced, which is much tighter in cases of coefficient blowup. This bound was relaxed slightly to keep the bounds/intervals in QQ.