Fix compilation error with bdwgc 7
The new garbage collector timing feature added to "time" relies on two
functions that were added in bdwgc 8.  They also might be unavailable
if bdwgc was compiled with NO_CLOCK.  In either case, compilation will
fail.  This is a problem, for example, on the Ubuntu LTS releases
18.04 and 20.04, which both have bdwgc 7.

We add workarounds for when these functions are unavailable.  Note
that if this is the case, then "time" will always print 0 for time
spent in garbage collection.