Simplify construction of empty matrices
Update map(Module, Nothing, List) so it behaves the same as
map(Module, Module, List) when the third argument is an empty list.
This lets us use "matrix(R, {})" to create a the empty matrix
representing the trivial map from the zero R-module to itself.

Also update matrix(List) so that "matrix {}" returns the empty matrix
with source and target ZZ^0.