Use "git describe" directly for git description (autotools)
Previously, we counted the number of commits since M2/VERSION was
modified.  Now we use "git describe" to count the number of commits
since the last release tag, as suggested by a TODO comment.

Also drop the "-unknown" from the git description when git isn't
present to match the cmake behavior.