Delete check for program licenses directory
It was left over from when "copyright" was a string, which is no
longer the case.

Also, the next thing after the "then" at the end of the line was the
definition of the "path" variable.  So if the "program licenses"
directory doesn't exist (e.g., in the Debian/Ubuntu package where we
don't vendor any 3rd party programs or libraries), then we immediately
got the following error:

/usr/share/Macaulay2/Core/startup.m2:583:16:(0): error: expected
argument 1 to be a list or sequence

This was from a call to "join", which expects lists and sequences, but
argument 1 was "path", which was still just a symbol.