Fix paths to Style files in html documentation
This partially reverts #3277, which fixed viewHelp when the Style
package was installed to the application directory.  However, it did
this at the cost of breaking all the other documentation installed to
the application directory!

The problem was that when calling makePackageIndex, the installTail
variable is set to an absolute path (the application directory), but
when building the package documentation, it is a relative path.  So we
can't use relativizeFilename in the same way in both situations.