Update PositivityToricBundles to v1.9
Added the method wellformedBundleFiltrations, updated documentation accordingly and added a corresponding test.
Several methods make the implicity assumption that the filtration steps are stored in an ascending manner. This is not true for bundles which are created by the methods dual or tensor (maybe others) from the package ToricVectorBundles.
The method wellformedBundleFiltrations ensures that the filtration steps are stored in a well-formed manner, suitable for the methods of this package.