Refactor contents of TestInput
Rather than storing the filename and line number separately, we store
a FilePosition containing both of them.  We update locate(TestInput)
to use this.

We also remove the redundant source file stamp from the code -- there's
no reason to also store the file location here.

Finally, we remove the NewFromMethod for constructing a TestInput from
a sequence.  There was really no reason to use this, as we only
construct a TestInput object once, inside TEST.