Drop "get" example from "runProgram" docs.
It's not directly related to runProgram.  Add a bit more detail about
how "get" might be used (by adding a leading "!"), and also link to
the docs for the various "open*" functions that also take a string
with a leading "!" to interact with a program.

The primary reason we're dropping this is example is that it called
gfan, which may not be on the PATH, especially if we've just built it.