Graphs: fix isCyclic Digraph
This fixes the detection of cycles in directed graphs (see issue #3626) by
not relying on the numbers from DFS but by attempting a topological sort.
It succeeds if and only if the graph is acyclic. This reuses existing code
as much as possible.

Also added tests that document previous bugs which are now fixed.