Fix --srcdir option when running examples/tests
We cd into a temporary directory when doing this, so if any --srcdir
options were given using relative paths, then we weren't looking in
the right place.

Previously, this was silently ignored, but now that we raise an error
when the Macaulay2 subdirectory doesn't exist, we were getting errors.

We fix this by relativizing the filename from the run directory to the
source directory.