Set googletest to version 1.16.0
We're currently using a commit a little between 1.16 and the
just-released 1.17.  googletest 1.17 (and the commit we're currently
using) requires compiling with C++17 support.  This is currently an
issue on systems like RHEL where we need to build googletest.

Eventually, we should add support for 1.17 by compiling with C++17
support in memtailor/mathic/mathicgb, but it would probably be best to
deal with this after the 1.25.05 release.

[ci skip]