Remove set_addrinfo helper function
It was a thin wrapper around getaddrinfo that used a global variable
to store the error message when an error was raised.  This likely
wasn't thread safe.

Instead, we determine the error message a bit later, when it's time
to actually print it, without any global variables.  This requires a
little bit of logic to distinguish between getaddrinfo errors and
non-getaddrinfo errors.  In particular, we shift the error codes of
the getaddrinfo errors and then shift back later when we need them.