Fix wait(List) so that it matches the documentation
The value of status will only be accurate if the process has changed
state, i.e., if waitpid returned a positive value.  So in particular,
we won't be able to trust that fix_status will return -2 in this case
as given in the documentation.  So we return -2 manually in this case.