Add option to initialize Python w/ a particular executable
This will allow us to use virtual environments

We're also initializing Python using an isolated configuration, which
is recommended since it's embedded inside another application.  This
means that we can't access the builtin functions using pythonValue, so
we need to update a bunch of code to get around this (e.g.,
pythonValue "foo" -> builtins@@"foo").