Add PythonRunScript to complement PythonValue
Both are wrappers around PyRun_String, but one evaluates a single line
and returns its output, and the other evaluates potentially many lines
and returns a dictionary containing its variables.

Also move all the C API code to python.d, and add a Global option so
we can pass a dictionary with variables.