Refactor "needs"
In particular, the unexported filesLoaded package has been updated so
that its keys are now absolute paths and its values are timestamps for
when the file was last modified.  Previously, the keys were relative
paths and the values were pairs consisting of absolute paths and
timestamps.

The relative path was used to pass to "load" so we could search in the
various directories in "path".  But we refactor "needs" so that it calls
the same helper function as "load" that does this searching for
us. This way, we'll be sure to load the file that appears first in the
directories in "path" as opposed to the one that just happened to be
stored in the value of the filesLoaded hash table.  This may not
have been accurate, e.g., if a new version of the file was created in
a directory earlier in "path" since the last time it was loaded.  In
particular, "needs" will now have the same behavior as "load" and
"input", as far as searching for files is concerned.