Delete old bug file related to parallel assignment
There was likely a typo in the file, and the last input likely was
supposed to be b and not a.

This apparently was "fixed" at some point, as this is the behavior in
Macaulay2 1.24.11:

i1 : (b) = 1:x

o1 = 1 : (x)

o1 : Sequence

i2 : b

o2 = x

o2 : Symbol

However, this seems incorrect, as (b) is not a sequence in Macaulay2,
but just b itself surrounded by parentheses for grouping.