revert the switch of choice of LU decomposition for RR_53, CC_53.
I had switched the LU decomp routine to use the builtin "naive" version, so that an LU decomp produces a row echelon matrix (not reduced though).  However, the numerics of that version mess up computations in some of the numerical algebraic geometry routines, where  `solve` then gives bad vectors, if any...  Reverting fixes that for now.  We still need to fix #2941 (nullSpace), and perhaps fix the naive version to have better numerical stability.