Add --enable-download option to check-configure
During configure, we raise an error if the input for --enable-download
is "yes" but neither wget nor curl are available on the system.

Now that the default for --enable-download is "yes", we should make
sure that we don't raise this error during check-configure if the user
initially ran configure with "no", e.g., in Debian/Ubuntu package builds.