Add support for new flint rand functions
In a recent version of flint, the flint_randinit and flint_randclear
functions were renamed to flint_rand_init and flint_rand_clear.  The
older versions are still supported, but result in deprecation warnings
and may be removed at some point.

We add FLINT_RAND_INIT and FLINT_RAND_CLEAR macros to cover both older
and newer versions of flint.